refactor(tts/g2p): share English phoneme helpers between Kokoro and StyleTTS2
Extract `stemInflected`, `spelledOutTokens`, and `letterPronunciations`
from `KokoroChunker` into a new `Sources/FluidAudio/TTS/G2P/EnglishG2PHelpers.swift`
so both backends use the same morphological-stemming + numeric-spell-out
+ per-letter fallback logic. The lexicon (`us_lexicon_cache.json`) is
already shared; the orchestration around it now is too.

Helpers are pure (no model loading, no I/O) and accept an optional
`allowed: Set<String>?` filter — `nil` accepts everything (StyleTTS2's
behavior, raw misaki phonemes), a non-nil set restricts to the Kokoro
178-token vocab.

KokoroChunker.swift: delete 224 lines of helpers (`spelledOutTokens`,
`letterPronunciations`, `usTaus`, `stemInflected`, `stemS`, `stemEd`,
`stemIng`, `appendSSuffix`, `appendEdSuffix`, `appendIngSuffix`) and the
unused `decimalDigits` static. Three call sites in `resolvePhonemes`
delegate to `EnglishG2PHelpers`. No behavioral change.

StyleTTS2Phonemizer.swift: extend `flushWordEnglish` from
lexicon → acronym → BART
to the full 6-step path
lexicon → acronym → stem → BART → spell-out → letterPronunciations.
The acronym path stays first (it uses the lexicon's case-sensitive
single-letter entries which carry stress markers — better than the
unstressed `letterPronunciations` table). LexiconHolder gains a
`lowerSnapshot()` accessor for the stemmer.

Smoke test "Hello world, this is a test of the running pipelines in 2025."
now produces:
  həlˈoʊ wˈɝld ðɪs ɪz eɪ tˈɛst ʌv ði ɹˈʌnɪŋ pˈaɪplˌaɪnz ɪn tˈu θˈaʊzənd twˈɛnti-fˈaɪv.
- "running" / "pipelines" resolved via stemInflected (plural / progressive).
- "2025" spelled out and resolved per-token via lexicon.

Tests:
- Tests/FluidAudioTests/TTS/G2P/EnglishG2PHelpersTests.swift — 11 tests
  pinning stem rules (-s, -ed, -ing, silent-e, doubled consonant, allowed
  filter), spell-out (basic, year, non-numeric reject), and full a-z
  letter table coverage.
- Existing StyleTTS2PhonemizerTests + 89 Kokoro suite tests all pass.
